chore: doc comments and cleanup
