+++ Cargo.toml
@@ -1,5 +1,7 @@
 [package]
 name = "libcamera-rs"
+description = "Libcamera bindings for rust."
+authors = ["John Brandt <johnb0@outlook.com>", "Ben Heard <benpheard@gmail.com>"]
 version = "0.1.0"
 edition = "2021"
