+++ readme.md
@@ -1,35 +1,41 @@
-use crate::prelude::{CameraEvent, CameraManager, PixelFormat, StreamRole};
+# Libcamera-rs
+[Libcamera](https://www.libcamera.org/) bindings for [Rust](https://www.rust-lang.org/).
 
-#[cfg(feature = "image")]
-#[test]
-fn test_camera() {
+## Example usage to take a bunch of pictures and change some camera parameters:
+```rust
+use libcamera_rs::prelude::*;
+
+fn main() {
   let mut cm = CameraManager::new().unwrap();
-  println!("cm: {cm:?}");
+  // Get the first camera from the camera manager.
   let mut cam = cm.get_camera_by_name(&cm.get_camera_names()[0]).unwrap();
-  println!("cam: {cam:?}");
+  // Generate a configuration with one "viewfinder" optimized stream.
   let conf = cam.generate_config(&[StreamRole::Viewfinder]).unwrap();
-  println!("conf: {conf:?}");
   let stream = &mut conf.streams_mut()[0];
+  // Attempt to assign the stream's pixel format and size.
   stream.set_pixel_format(PixelFormat::Mjpeg);
   stream.set_size(640, 480);
-  println!("Configuration applied: {:?}", cam.apply_config().unwrap());
+  // Start the camera stream.
   cam.start_stream().unwrap();
-  println!("Started stream.");
-  println!("Capturing frames...");
-  for i in 0..6 {
+  // Take 10 pictures
+  for i in 0..10 {
+    // Change brightness and contrast
     let brightness = cam.get_controls_mut().brightness.as_mut().unwrap();
-    brightness.set_value((brightness.get_value() + 0.6).rem_euclid(1.5) - 0.5);
+    brightness.set_value((brightness.get_value() + 0.9).rem_euclid(1.5) - 0.5);
     let contrast = cam.get_controls_mut().contrast.as_mut().unwrap();
-    contrast.set_value(contrast.get_value() + 0.02);
+    contrast.set_value(contrast.get_value() + 0.1);
+    // Queue the capture request
     cam.capture_next_picture(0).unwrap();
-    println!("Capturing image #{}", i);
+    // Wait for a bit.
     std::thread::sleep(std::time::Duration::from_millis(200));
+    // Poll events (containing the images)
     let events = cam.poll_events().unwrap();
     for event in events {
       match event {
         CameraEvent::RequestComplete {
           serial_id, image, ..
         } => {
+          // Reencode the image to PNG and save it.
           let decoded_image = image.try_decode().unwrap();
           let rgb_image = decoded_image
             .as_bgr()
@@ -39,10 +45,11 @@ fn test_camera() {
             .unwrap();
           let filename = format!("image_{serial_id}.png");
           std::fs::write(&filename, rgb_image).unwrap();
-          println!("Got responce back for request {serial_id} and saved to {filename}.");
         }
       }
     }
   }
-  println!("Done!");
 }
+```
+
+Most functions should be well enough documented with rustdoc.
