+++ src/lib.rs
@@ -1,34 +1,20 @@
-use thiserror::Error;
+#![warn(missing_docs)]
+#![doc = include_str!("../readme.md")]
 
 mod bridge;
+/// Handles interfacing with cameras, get started using a CameraManager.
 pub mod camera;
+/// Camera and stream configuration, e.g. how many streams, resolutions, and pixel formats.
 pub mod config;
+/// Camera (runtime) controls, e.g. brightness, contrast.
 pub mod controls;
+/// Errors
+pub mod error;
+/// Decoding images from the camera
 pub mod image;
+/// All the things you *should* need to get started.
 pub mod prelude;
 #[cfg(test)]
 mod test;
 
-#[derive(Debug, Error)]
-pub enum LibcameraError {
-  #[error("Inner C++ error: {0}")]
-  InnerError(#[from] cxx::Exception),
-  #[error("Int conversion error: {0}")]
-  IntConversion(#[from] std::num::TryFromIntError),
-  #[error("Configuration Invalid or Missing")]
-  InvalidConfig,
-  #[error("No buffer ready for capture (all buffers in use, capture pictures slower!)")]
-  NoBufferReady,
-  #[error("Control value out of range!")]
-  InvalidControlValue,
-  #[error("Unknown ID in camera request")]
-  UnknownRequestId,
-  #[cfg(feature = "image")]
-  #[error("Image error: {0}")]
-  ImageError(#[from] ::image::ImageError),
-  #[cfg(feature = "image")]
-  #[error("Image error: Bad image format.")]
-  BadImageFormat,
-}
-
-type Result<T> = std::result::Result<T, LibcameraError>;
+pub use error::{LibcameraError, Result};
